feat: dev브랜치와 병합 & 폴더구조 수정
