+++ src/components/My/index.module.scss
@@ -1,26 +1,24 @@
-.myPageContainer {
+.container {
   display: flex;
   flex-direction: column;
   row-gap: 20px;
   margin: 0 20px;
 
+  a {
+    text-decoration: none;
+    color: inherit;
+  }
+
   button {
-    margin: 0;
-    padding: 0;
     border: none;
-    background-color: transparent;
-    cursor: pointer;
+    background-color: none;
   }
 
-  li {
-    list-style: none;
-  }
-
-  .myPageTit {
+  .tit {
     margin: 0;
   }
 
-  .myPageContents {
+  .contents {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
@@ -30,7 +28,7 @@
     border-radius: 8px;
     background-color: #ddd;
 
-    .myPageMyImg {
+    .myImg {
       display: flex;
       justify-content: center;
       align-items: center;
@@ -42,19 +40,19 @@
       span {}
     }
 
-    .myPageInfoWrap {
+    .infoWrap {
       display: flex;
       flex-direction: column;
       align-items: center;
 
-      .myPageMyNickname {}
+      .nickname {}
 
-      .myPageBadgeNotice {
+      .badgeNotice {
         strong {}
       }
     }
 
-    .myPageStatWrap {
+    .statWrap {
       display: flex;
       column-gap: 40px;
 
@@ -71,7 +69,7 @@
 
   }
 
-  .myPageGetBadgeStatus {
+  .myBadgeStatus {
     display: flex;
     align-items: center;
     column-gap: 20px;
@@ -79,7 +77,7 @@
     border-radius: 10px;
     padding: 20px;
 
-    .myPageBadgeImg {
+    .badgeImg {
       width: 50px;
       height: 50px;
       border-radius: 50%;
@@ -89,7 +87,7 @@
     span {}
   }
 
-  .myPageShareBtn {
+  .shareBtn {
     margin: 0 auto;
     padding: 20px;
     background-color: pink;
